Combining General and TestSetup into one class, as remaining functionality is all similar
