+++ src/main/java/com/coveros/selenified/selenium/WaitFor.java
@@ -26,7 +26,6 @@ import org.testng.log4testng.Logger;
 
 import com.coveros.selenified.OutputFile;
 import com.coveros.selenified.OutputFile.Result;
-import com.coveros.selenified.utilities.General;
 
 /**
  * WaitFor performs dynamic waits on the app in general, until a particular
@@ -41,7 +40,7 @@ import com.coveros.selenified.utilities.General;
  */
 public class WaitFor {
 
-    private static final Logger log = Logger.getLogger(General.class);
+    private static final Logger log = Logger.getLogger(WaitFor.class);
 
     // this will be the name of the file we write all commands out to
     private OutputFile file;
