+++ src/main/java/com/coveros/selenified/selenium/element/WaitFor.java
@@ -28,7 +28,6 @@ import org.testng.log4testng.Logger;
 import com.coveros.selenified.OutputFile;
 import com.coveros.selenified.OutputFile.Result;
 import com.coveros.selenified.selenium.element.Element;
-import com.coveros.selenified.utilities.General;
 
 /**
  * WaitFor performs dynamic waits on a particular element, until a particular
@@ -42,7 +41,7 @@ import com.coveros.selenified.utilities.General;
  */
 public class WaitFor {
 
-    private static final Logger log = Logger.getLogger(General.class);
+    private static final Logger log = Logger.getLogger(WaitFor.class);
 
     // what element are we trying to interact with on the page
     private Element element;
