+++ src/main/java/com/coveros/selenified/utilities/TestSetup.java
@@ -20,9 +20,13 @@
 
 package com.coveros.selenified.utilities;
 
+import java.lang.reflect.Method;
 import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.Random;
+
 import org.openqa.selenium.Proxy;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.chrome.ChromeDriver;
@@ -54,7 +58,7 @@ import io.github.bonigarcia.wdm.OperaDriverManager;
  */
 public class TestSetup {
 
-    private static final Logger log = Logger.getLogger(General.class);
+    private static final Logger log = Logger.getLogger(TestSetup.class);
 
     // constants
     private static final String PROXY_INPUT = "proxy";
@@ -64,6 +68,9 @@ public class TestSetup {
     private static final String DEVICE_NAME_INPUT = "deviceName";
     private static final String DEVICE_ORIENTATION_INPUT = "deviceOrientation";
     private static final String DEVICE_PLATFORM_INPUT = "devicePlatform";
+
+    private static final int MAXFILENAMELENGTH = 200;
+    private static final String PUBLIC = "public";
 
     private DesiredCapabilities capabilities;
 
@@ -139,7 +146,7 @@ public class TestSetup {
         } else {
             String[] allDetails = System.getProperty(BROWSER_INPUT).split(",");
             for (String details : allDetails) {
-                Map<String, String> browserDetails = General.parseMap(details);
+                Map<String, String> browserDetails = parseMap(details);
                 if (browserDetails.containsKey(BROWSER_NAME_INPUT)) {
                     browsers.add(Browser.lookup(browserDetails.get(BROWSER_NAME_INPUT)));
                 } else {
@@ -294,5 +301,187 @@ public class TestSetup {
             throw new InvalidBrowserException("The selected browser " + browser + " is not an applicable choice");
         }
         return driver;
+    }
+
+    /**
+     * Generates a random string of alpha-numeric characters
+     * 
+     * @param length
+     *            the length of the random string
+     * @return String: random string of characters
+     */
+    public static String getRandomString(int length) {
+        if (length <= 0) {
+            return "";
+        }
+        String stringChars = "0123456789ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz";
+        Random rnd = new Random();
+        StringBuilder sb = new StringBuilder(length);
+        for (int i = 0; i < length; i++) {
+            sb.append(stringChars.charAt(rnd.nextInt(stringChars.length())));
+        }
+        return sb.toString();
+    }
+
+    /**
+     * Removes all non alphanumeric characters from a provided string
+     * 
+     * @param value
+     *            - the string to cleanup
+     * @return String: the provided string with all alphanumeric characters
+     *         removed
+     */
+    public static String removeNonWordCharacters(String value) {
+        if (value == null) {
+            return value;
+        }
+        return value.replaceAll("[^a-zA-Z0-9]+", "");
+    }
+
+    /**
+     * Determines the unique test name, based on the parameters passed in
+     * 
+     * @param method
+     *            - the method under test to extract the name from
+     * @param dataProvider
+     *            - an array of objects being passed to the test as data
+     *            providers
+     * @return String: a unique name
+     */
+    public static String getTestName(Method method, Object... dataProvider) {
+        String className;
+        String packageName = "";
+        if (method.getDeclaringClass().getPackage() != null) {
+            className = method.getDeclaringClass().toString().substring(6).split("\\.")[1];
+            packageName = method.getDeclaringClass().toString().substring(6).split("\\.")[0];
+        } else {
+            className = method.getDeclaringClass().toString().substring(6);
+        }
+        return getTestName(packageName, className, method.getName(), dataProvider);
+    }
+
+    /**
+     * Determines if a dataProvider was actually provided, or just ITestContext
+     * or method data is present
+     * 
+     * @param dataProvider
+     *            - the object array to check - is it truly a data provider?
+     * @return Boolean: is the provided object array a data provider?
+     */
+    private static boolean isRealDataProvider(Object... dataProvider) {
+        return dataProvider != null && dataProvider.length > 0 && dataProvider[0] != null
+                && !dataProvider[0].toString().startsWith(PUBLIC);
+    }
+
+    /**
+     * Determines the unique test name, based on the parameters passed in
+     * 
+     * @param packageName
+     *            - the package name of the test method as a string
+     * @param className
+     *            - the class name of the test method as a string
+     * @param methodName
+     *            - the method name of the test as a string
+     * @param dataProvider
+     *            - an array of objects being passed to the test as data
+     *            providers
+     * @return String: a unique name
+     */
+    public static String getTestName(String packageName, String className, String methodName, Object... dataProvider) {
+        StringBuilder testName;
+        if ("".equals(packageName)) {
+            testName = new StringBuilder(className + "_" + methodName);
+        } else {
+            testName = new StringBuilder(packageName + "_" + className + "_" + methodName);
+        }
+        String currentName = testName.toString();
+        if (isRealDataProvider(dataProvider)) {
+            testName.append("WithOption");
+            for (Object data : dataProvider) {
+                if (data == null || data.toString().startsWith(PUBLIC)) {
+                    break;
+                }
+                testName.append(TestSetup.capitalizeFirstLetters(removeNonWordCharacters(data.toString())));
+            }
+            currentName = testName.toString();
+            if (currentName.length() > MAXFILENAMELENGTH) {
+                if ("".equals(packageName)) {
+                    currentName = className + "_" + methodName + dataProvider.toString().split(";")[1]; // NOSONAR
+                                                                                                        // -
+                                                                                                        // purposefully
+                                                                                                        // using
+                                                                                                        // toString
+                                                                                                        // on
+                                                                                                        // object
+                                                                                                        // to
+                                                                                                        // obtain
+                                                                                                        // unique
+                                                                                                        // random
+                                                                                                        // hash
+                } else {
+                    currentName = packageName + "_" + className + "_" + methodName
+                            + dataProvider.toString().split(";")[1]; // NOSONAR
+                                                                        // -
+                                                                        // purposefully
+                                                                        // using
+                                                                        // toString
+                                                                        // on
+                                                                        // object
+                                                                        // to
+                                                                        // obtain
+                                                                        // unique
+                                                                        // random
+                                                                        // hash
+                }
+            }
+        }
+        return currentName;
+    }
+
+    /**
+     * Capitalizes the first letter of each word in the provided string
+     * 
+     * @param word
+     *            - the string to be capitalized on
+     * @return String: the new string
+     */
+    public static String capitalizeFirstLetters(String word) {
+        if (word == null) {
+            return word;
+        }
+        String out = "";
+        for (int i = 0; i < word.length(); i++) {
+            if (i == 0) {
+                // Capitalize the first letter of the string.
+                out = String.format("%s%s", Character.toUpperCase(word.charAt(0)), word.substring(1));
+            }
+            // Is this character a non-letter? If so
+            // then this is probably a word boundary so let's capitalize
+            // the next character in the sequence.
+            if (!Character.isLetter(out.charAt(i)) && (i + 1) < out.length()) {
+                out = String.format("%s%s%s", out.subSequence(0, i + 1), Character.toUpperCase(out.charAt(i + 1)),
+                        out.substring(i + 2));
+            }
+        }
+        return out;
+    }
+
+    /**
+     * Breaks up a string, and places it into a map. ampersands (&) are used to
+     * split into key value pairs, while equals (=) are used to assign key vs
+     * values
+     * 
+     * @param input
+     *            - a string, with key and values separated by an equals (=) and
+     *            pairs separated by an ampersand (&)
+     * @return Map: a map with values
+     */
+    public static Map<String, String> parseMap(final String input) {
+        final Map<String, String> map = new HashMap<>();
+        for (String pair : input.split("&")) {
+            String[] kv = pair.split("=");
+            map.put(kv[0], kv[1]);
+        }
+        return map;
     }
 }
